refactor: BoardDTO와 PlayerDTO 각 도메인 패키지로 이동 및 Board의 getter 접근 제어자 default로 변경
