+++ src/main/java/janggi/JanggiGame.java
@@ -4,8 +4,8 @@ import janggi.domain.board.Board;
 import janggi.domain.game.Players;
 import janggi.domain.board.Position;
 import janggi.domain.game.Side;
-import janggi.dto.BoardDTO;
-import janggi.dto.PlayerDTO;
+import janggi.domain.board.BoardDTO;
+import janggi.domain.game.PlayerDTO;
 import janggi.view.InputView;
 import janggi.view.OutputView;
 import java.util.HashSet;
