+++ src/main/java/janggi/domain/game/PlayerDTO.java
@@ -1,6 +1,4 @@
-package janggi.dto;
-
-import janggi.domain.game.Side;
+package janggi.domain.game;
 
 public record PlayerDTO(String name, Side side) {
 }
